listening-comp/backend/requirements.txt
chromadb==0.4.18
streamlit
boto3
youtube_transcript_api
python-dotenv==1.0.0
requests==2.31.0
openai==1.3.7
pathlib==1.0.1